feat(plugin-native-navigation): updated package to version 8 to support newArch
